Dissect visitor.php
