+++ phpcs.xml.dist
@@ -1,8 +1,8 @@
 <?xml version="1.0"?>
 <ruleset>
     <file>finder.php</file>
-    <file>functionMap.php</file>
     <file>visitor.php</file>
+    <file>src/</file>
 
     <rule ref="PSR12NeutronRuleset">
         <exclude name="Generic.Files.LineLength"/>
